requirements.txt
fastapi
fastapi-sqlalchemy
pydantic
alembic
psycopg2
uvicorn
python-dotenv
pandas
Pillow>=8.3.1
Jinja2==3.0.3
python-multipart
tensorflow
h5py
google-cloud-bigquery
db-dtypes
google-cloud-storage
